+++ packages/core/src/__tests__/cogniteClient.unit.spec.ts
@@ -3,7 +3,7 @@
 import nock from 'nock';
 import BaseCogniteClient from '../baseCogniteClient';
 
-import { API_KEY_HEADER, AUTHORIZATION_HEADER, BASE_URL } from '../constants';
+import { API_KEY_HEADER, BASE_URL } from '../constants';
 import { apiKey, project } from '../testUtils';
 
 const mockBaseUrl = 'https://example.com';
@@ -12,9 +12,18 @@ function setupClient(baseUrl: string = BASE_URL) {
   return new BaseCogniteClient({
     appId: 'JS SDK integration tests',
     project: 'test-project',
-    getToken: () => Promise.resolve(apiKey),
-    apiKeyMode: true,
     baseUrl,
+    apiKeyMode: true,
+    getToken: () => Promise.resolve(apiKey),
+  });
+}
+
+function setupNoAuthClient(noAuthMode = true) {
+  return new BaseCogniteClient({
+    appId: 'JS SDK integration tests',
+    project: 'test-project',
+    baseUrl: mockBaseUrl,
+    noAuthMode: noAuthMode,
   });
 }
 
@@ -65,24 +74,25 @@ describe('CogniteClient', () => {
       );
     });
 
-    describe('getToken', () => {
-      test('missing getToken', () => {
+    test('not throw errors when running in noAuthMode', () => {
+      expect(() => {
+        // @ts-ignore
+        setupNoAuthClient(true);
+      }).not.toThrow();
+    });
+
+    describe('credentials', () => {
+      test('missing credentials', () => {
         expect(() => {
           // @ts-ignore
           new BaseCogniteClient({ appId: 'unit-test', project: 'unit-test' });
         }).toThrowErrorMatchingInlineSnapshot(
-          `"options.getToken is required and must be of type () => Promise<string>"`
+          `"options.authentication.credentials is required or options.getToken is request and must be of type () => Promise<string>"`
         );
       });
-      test('call getToken on 401', async () => {
-        const getToken = jest.fn().mockResolvedValue('401-test-token');
-
-        nock(mockBaseUrl).get('/test').reply(401, {});
-        nock(mockBaseUrl, {
-          reqheaders: {
-            [AUTHORIZATION_HEADER]: 'Bearer 401-test-token',
-          },
-        })
+      test('call credentials on 401', async () => {
+        nock(mockBaseUrl)
+          .persist()
           .get('/test')
           .reply(200, { body: 'request ok' });
 
@@ -90,28 +100,28 @@ describe('CogniteClient', () => {
           project,
           appId: 'unit-test',
           baseUrl: mockBaseUrl,
-          getToken,
+          apiKeyMode: true,
+          getToken: () => Promise.resolve('401-test-token'),
         });
 
         const result = await client.get('/test');
 
-        expect(getToken).toHaveBeenCalledTimes(1);
         expect(result.status).toEqual(200);
         expect(result.data).toEqual({ body: 'request ok' });
+
+        // scopeOk.done();
       });
 
       test('getToken rejection should reject sdk requests', async () => {
         const getToken = jest.fn().mockRejectedValue(new Error('auth error'));
 
-        nock(mockBaseUrl)
-          .get('/test')
-
-          .reply(401, {});
+        nock(mockBaseUrl).get('/test').reply(401, {});
 
         const client = new BaseCogniteClient({
           project,
           appId: 'unit-test',
           baseUrl: mockBaseUrl,
+          apiKeyMode: true,
           getToken,
         });
 
@@ -126,36 +136,18 @@ describe('CogniteClient', () => {
     });
 
     test('apiKeyMode should change request header and token preable', async () => {
-      const getToken = jest.fn().mockResolvedValue('test-api-key');
-
-      nock(mockBaseUrl).get('/test').reply(401, {});
-      nock(mockBaseUrl, {
-        reqheaders: {
-          [API_KEY_HEADER]: 'test-api-key',
-        },
-      })
-        .get('/test')
-        .reply(200, { body: 'api key request ok' });
-
-      nock(mockBaseUrl, {
-        reqheaders: {
-          [AUTHORIZATION_HEADER]: 'Bearer 401-test-token',
-        },
-      })
-        .get('/test')
-        .reply(200, { body: 'normal token request ok' });
+      nock(mockBaseUrl).get('/test').reply(200, { body: 'api key request ok' });
 
       const client = new BaseCogniteClient({
         project,
         appId: 'unit-test',
         baseUrl: mockBaseUrl,
-        getToken,
         apiKeyMode: true,
+        getToken: () => Promise.resolve('test-api-key'),
       });
 
       const result = await client.get('/test');
 
-      expect(getToken).toHaveBeenCalledTimes(1);
       expect(result.status).toEqual(200);
       expect(result.data).toEqual({ body: 'api key request ok' });
     });
@@ -169,11 +161,45 @@ describe('CogniteClient', () => {
     nock(mockBaseUrl, { reqheaders: expectedHeaders }).get('/').reply(200, {});
   });
 
-  describe('http requests', () => {
+  describe('authenticated http requests', () => {
     let client: BaseCogniteClient;
 
     beforeAll(async () => {
       client = setupMockableClient();
+    });
+
+    test('get method', async () => {
+      nock(mockBaseUrl).get('/').once().reply(200, []);
+      const response = await client.get('/');
+      expect(response.data).toEqual([]);
+    });
+
+    test('post method', async () => {
+      nock(mockBaseUrl).post('/').once().reply(200, []);
+      const response = await client.post('/');
+      expect(response.data).toEqual([]);
+    });
+
+    test('put method', async () => {
+      nock(mockBaseUrl).put('/').once().reply(200, []);
+      const response = await client.put('/', {
+        responseType: 'json',
+      });
+      expect(response.data).toEqual([]);
+    });
+
+    test('delete method', async () => {
+      nock(mockBaseUrl).delete('/').once().reply(200, 'abc');
+      const response = await client.delete('/', { responseType: 'text' });
+      expect(response.data).toBe('abc');
+    });
+  });
+
+  describe('noAuthMode http requests', () => {
+    let client: BaseCogniteClient;
+
+    beforeAll(async () => {
+      client = setupNoAuthClient();
     });
 
     test('get method', async () => {