+++ packages/core/CHANGELOG.md
@@ -3,6 +3,56 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.5.1](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-core@4.5.0...@cognite/sdk-core@4.5.1) (2022-07-07)
+
+
+### Bug Fixes
+
+* return token instead of bearertoken ([#834](https://github.com/cognitedata/cognite-sdk-js/issues/834)) ([2829149](https://github.com/cognitedata/cognite-sdk-js/commit/2829149c906b1ecd2c1f28808738ddc28f8f798c))
+
+
+
+
+
+# [4.5.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-core@4.4.0...@cognite/sdk-core@4.5.0) (2022-06-17)
+
+
+### Features
+
+* noAuthMode support ([#824](https://github.com/cognitedata/cognite-sdk-js/issues/824)) ([69a33ce](https://github.com/cognitedata/cognite-sdk-js/commit/69a33ce1f352706a9e36b119a99be9132e9e3c9c))
+
+
+
+
+
+# [4.4.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-core@4.3.0...@cognite/sdk-core@4.4.0) (2022-06-09)
+
+
+### Features
+
+* auth wrapper as a provider ([#816](https://github.com/cognitedata/cognite-sdk-js/issues/816)) ([4c2fc03](https://github.com/cognitedata/cognite-sdk-js/commit/4c2fc0367e1aba347c6d8ab0a3b1e5f3a15a7a2d))
+
+
+
+
+
+# [4.3.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-core@4.2.0...@cognite/sdk-core@4.3.0) (2022-06-02)
+
+
+### Features
+
+* automatic token management with @cognite/auth-wrapper  ([#796](https://github.com/cognitedata/cognite-sdk-js/issues/796)) ([d13c53f](https://github.com/cognitedata/cognite-sdk-js/commit/d13c53f9d5a5e65d6e2fc7d1a0e2efe2d36e678c))
+* automatic token management with @cognite/auth-wrapper  ([#814](https://github.com/cognitedata/cognite-sdk-js/issues/814)) ([d2c3a30](https://github.com/cognitedata/cognite-sdk-js/commit/d2c3a3083746e234f5dfc3e290188414d9ea17a6))
+
+
+### Reverts
+
+* Revert "feat: automatic token management with @cognite/auth-wrapper  (#796)" (#813) ([0a8a61c](https://github.com/cognitedata/cognite-sdk-js/commit/0a8a61c2b2c3d992f80b40f6282b20595537edac)), closes [#796](https://github.com/cognitedata/cognite-sdk-js/issues/796) [#813](https://github.com/cognitedata/cognite-sdk-js/issues/813)
+
+
+
+
+
 # [4.2.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-core@4.1.7...@cognite/sdk-core@4.2.0) (2022-05-27)