+++ packages/playground/CHANGELOG.md
@@ -3,6 +3,68 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.5.1](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-playground@5.5.0...@cognite/sdk-playground@5.5.1) (2022-07-07)
+
+**Note:** Version bump only for package @cognite/sdk-playground
+
+
+
+
+
+# [5.5.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-playground@5.4.1...@cognite/sdk-playground@5.5.0) (2022-06-20)
+
+
+### Features
+
+* **annotations:** move annotations to stable ([#823](https://github.com/cognitedata/cognite-sdk-js/issues/823)) ([b453685](https://github.com/cognitedata/cognite-sdk-js/commit/b4536855ad62e9888af11419b89af1621a50a7ff))
+
+
+
+
+
+## [5.4.1](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-playground@5.4.0...@cognite/sdk-playground@5.4.1) (2022-06-17)
+
+**Note:** Version bump only for package @cognite/sdk-playground
+
+
+
+
+
+# [5.4.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-playground@5.3.1...@cognite/sdk-playground@5.4.0) (2022-06-13)
+
+
+### Features
+
+* **annotations:** remove linked resource fields ([#822](https://github.com/cognitedata/cognite-sdk-js/issues/822)) ([248d2a0](https://github.com/cognitedata/cognite-sdk-js/commit/248d2a0ba3ecbd3625e20e26bb7cd051f276febe))
+
+
+
+
+
+## [5.3.1](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-playground@5.3.0...@cognite/sdk-playground@5.3.1) (2022-06-09)
+
+**Note:** Version bump only for package @cognite/sdk-playground
+
+
+
+
+
+# [5.3.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-playground@5.2.3...@cognite/sdk-playground@5.3.0) (2022-06-02)
+
+
+### Features
+
+* automatic token management with @cognite/auth-wrapper  ([#796](https://github.com/cognitedata/cognite-sdk-js/issues/796)) ([d13c53f](https://github.com/cognitedata/cognite-sdk-js/commit/d13c53f9d5a5e65d6e2fc7d1a0e2efe2d36e678c))
+
+
+### Reverts
+
+* Revert "feat: automatic token management with @cognite/auth-wrapper  (#796)" (#813) ([0a8a61c](https://github.com/cognitedata/cognite-sdk-js/commit/0a8a61c2b2c3d992f80b40f6282b20595537edac)), closes [#796](https://github.com/cognitedata/cognite-sdk-js/issues/796) [#813](https://github.com/cognitedata/cognite-sdk-js/issues/813)
+
+
+
+
+
 ## [5.2.3](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-playground@5.2.2...@cognite/sdk-playground@5.2.3) (2022-05-27)
 
 **Note:** Version bump only for package @cognite/sdk-playground